% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_functions.R
\name{add_upi_info}
\alias{add_upi_info}
\title{Add education details (UPI, nc_year_actual) to extracted data.}
\usage{
add_upi_info(extracted_data, KEY = "", verbose = T)
}
\arguments{
\item{extracted_data}{extracted data}

\item{verbose}{Flag (TRUE/FALSE) for console messages.}

\item{Key}{API token}
}
\value{
extracted data with education details appended.
}
\description{
Add education details (UPI, nc_year_actual) to extracted data.
}
\details{
This function should be consumed into get_primary_school_student_data() and get_secondary_school_student_data(). However, to not have to re-extract all the data again this was currently added as a separate function.
}
