% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_name}
\alias{clean_name}
\title{Clean names}
\usage{
clean_name(x, rm_whit = F)
}
\arguments{
\item{x}{name to clean}

\item{rm_whit}{Flag (TRUE/FALSE) for whether whitespace wants to be removed from the names}
}
\value{
cleaned name
}
\description{
Clean names
}
\details{
"Cleans" name by changing to lower case and translating to Latin-ASCII (i.e removing accents, etc). Can also remove whitespace by setting rm_whit = T.
}
\examples{
 clean_name('John Leonard Knapp')
 clean_name('Pierre-Joseph Redouté')

}
\seealso{
Other utils: 
\code{\link{combine_raw_data}()}
}
\concept{utils}
