% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rbind_aggro}
\alias{rbind_aggro}
\title{rbind dataframes with non-identical columns}
\usage{
rbind_aggro(list_of_data.frames)
}
\arguments{
\item{list_of_data.frames}{list of data frames}
}
\value{
rbinded data frame
}
\description{
rbind dataframes with non-identical columns
}
\details{
This function finds all unique column names across all data frames in `list_of_data.frames`
and in each data frame in the list adjoins the missing columns whose values are set to NA. Then rbind is applied to the list of data frames.
}
