% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{teacher_master_list_example}
\alias{teacher_master_list_example}
\title{Example class list dataset}
\format{
A data frame with 6 rows and 13 variables
}
\usage{
teacher_master_list_example
}
\description{
An example of a teacher master list containing 6 teachers within two schools and three classes. The variables are as follows:
}
\keyword{teacher_master_list}
