% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token_management.R
\name{wonde_use_token}
\alias{wonde_use_token}
\title{Helper to get and save Wonde API token}
\usage{
wonde_use_token()
}
\description{
Provides instructions on how to obtain and store an API token for the
Wonde API. Users must complete this step before using any functionality
that requires authentication. Obtaining a token requires contacting
the OME team via email.
}
\details{
After requesting access, you will receive an API token by email.
Once received, the token should be stored in your \file{.Renviron}
file under the name \env{WONDE_TOKEN} and R must be restarted for
the change to take effect.
}
\examples{
\dontrun{
# Display instructions to set up an Wonde API token
wonde_use_token()
}
}
