% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withdrawing.R
\name{apply_withdrawn}
\alias{apply_withdrawn}
\title{Withdraw students from a class list}
\usage{
apply_withdrawn(
  class_list,
  withdrawn_students =
    "Primary Cohort Study - Parent_guardian withdrawal form(1-89).xlsx",
  school_link = "~/Wonde/Establishment_List_ID.xlsx",
  links_category = "Primary",
  expected_years = c(2019, 2020),
  do_log = F
)
}
\arguments{
\item{class_list}{class list}

\item{withdrawn_students}{path to an .xlsx file containing the withdrawn students.}

\item{do_log}{}
}
\description{
Withdraw students from a class list
}
