% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{class_list_example}
\alias{class_list_example}
\title{Example class list dataset}
\format{
A data frame with 60 rows and 9 variables
}
\usage{
class_list_example
}
\description{
An example of a cleaned class list containing two schools and three classes. The variables are as follows:
\itemize{
  \item School name.
  \item URN.
  \item Class Name.
  \item Main Teacher First Name.
  \item Main Teacher Last Name.
  \item Pupil First Name.
  \item Pupil Last Name.
  \item DoB. Formatted YYYY-MM-DD.
  \item UPN.
}
}
\keyword{class_list}
