% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combine_raw_data}
\alias{combine_raw_data}
\title{combine two lists element-wise}
\usage{
combine_raw_data(a, b)
}
\arguments{
\item{a}{list}

\item{b}{list}
}
\value{
a and b combined into one list (i,e a[[1]], a[[2]], ... a[[N]], b[[1]], b[[2]], ... b[[M]].)
}
\description{
combine two lists element-wise
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
list1 = list(A=1, B=2, C=3)
list2 = list(D=1, E=2, F=3)
combine_raw_data(list1, list2)
\dontshow{\}) # examplesIf}
}
