% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_list_element_to_df}
\alias{convert_list_element_to_df}
\title{convert sublist element of data frame to a data frame}
\usage{
convert_list_element_to_df(data, column_to_unnest)
}
\arguments{
\item{data}{data}

\item{column_to_unnest}{column_to_unnest}
}
\description{
convert sublist element of data frame to a data frame
}
