% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_class_list_changes.R
\name{create_school_blanks}
\alias{create_school_blanks}
\title{Create a "Blank" class list given school and class information}
\usage{
create_school_blanks(
  schools_with_size,
  numbering_type = "school",
  teacher_name = c("Reception", "Teacher"),
  class_name = "Reception Class",
  type = 30,
  class_buffer = 5,
  class_size = 30,
  class_list_columns = "master"
)
}
\arguments{
\item{schools_with_size}{A data frame where each row consists of a school with columns
SchoolName, URN, classInfo (Optional), TeacherInfo (Optional), NoStudents (Pptional)..}

\item{teacher_name}{default teacher name in none are provided.}

\item{class_name}{default class name if none are provided}

\item{type}{type}

\item{class_buffer}{class buffer}

\item{class_size}{default class size if not provided.}

\item{class_list_columns}{either 'master', 'print' or a vector of the columns wanted in order.}
}
\value{
formatted class list for "blanked" schools
}
\description{
Create a "Blank" class list given school and class information
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
## Create some blanked schools.
A = c('Narcissus Primary School', 'Tulipa Primary School')
B = c('321123', '567765')
C = c('Pseudonarcissus -> 10', 'Albanica -> 30, Persica -> 29')
D = c('Carl---Linnaeus','Kit---Tan, Robert---Sweet ')
sch_info = data.frame(SchoolName = A,
                      URN = B,
                      ClassInfo = C,
                      TeacherInfo = D)
create_school_blanks(sch_info)

sch_info = data.frame(SchoolName = A,
                      URN = B,
                      NoStudents = c(25, 60))
out = create_school_blanks(sch_info)
\dontshow{\}) # examplesIf}
}
