% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_functions.R
\name{get_primary_school_student_data}
\alias{get_primary_school_student_data}
\title{Get school information}
\usage{
get_primary_school_student_data(school_id, KEY = "")
}
\arguments{
\item{school_id}{wonde school ID}

\item{KEY}{API token}
}
\description{
Get school information
}
\details{
Given the school ID and KEY the function queries the Wonde API to extract:

- student information
- student group information
- student class information
- staff information
- subject information (secondary only)

This is returned as a list of data frames containing each subset of information.

Primary school information is obtained using `get_primary_school_student_data()` and secondary school information from `get_secondary_school_student_data()`.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
school_id = "Enter_ID_here"
KEY = "Enter_KEY_here"
primary_school_ids = c('A', 'B', 'C', 'D')

# Get info for individual school.
primary_info = get_primary_school_student_data(school_id, KEY)
secondary_info = get_secondary_school_student_data(school_id, KEY)

# Loop over and extract info for all primary schools.
primary_data_raw = lapply(primary_school_ids, function(id){get_primary_school_student_data(id, KEY) })
names(primary_data_raw) = primary_school_ids
primary_data_raw
\dontshow{\}) # examplesIf}
}
