% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_functions.R
\name{get_query}
\alias{get_query}
\title{Pull from wonde}
\usage{
get_query(url_path, filter = "", KEY = "")
}
\arguments{
\item{url_path}{the API endpoint}

\item{filter}{Addition objects to pass via include.}

\item{KEY}{API token}
}
\value{
data frame of pulled information
}
\description{
Pull from wonde
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
school_id = "Enter_ID_here"
Key = "Enter_KEY_here"

# Employee information.
 url_path = paste0("https://api.wonde.com/v1.0/schools/", school_id, "/employees/")
 employees = get_query(url_path, filter = '&include=groups', KEY = KEY)
 employees_groups = employees |> convert_list_element_to_df(column_to_unnest = 'groups.data') |>
 tidyr::unnest(groups.data, names_sep ='__', keep_empty  = T)
\dontshow{\}) # examplesIf}
}
